renamed Helper folder -> Extensions & CompressionHelper namespace fixed
