+++ Exilence/Extensions/DistributedCacheExtensions.cs
